+++ lib/api-helper/machine/handlerRegistrations.ts
@@ -53,6 +53,7 @@ import { isProject } from "@atomist/automation-client/lib/project/Project";
 import { toFactory } from "@atomist/automation-client/lib/util/constructionUtils";
 import { GitHubRepoTargets } from "../../api/command/target/GitHubRepoTargets";
 import { isTransformModeSuggestion } from "../../api/command/target/TransformModeSuggestion";
+import { SdmContext } from "../../api/context/SdmContext";
 import { CommandListenerInvocation } from "../../api/listener/CommandListener";
 import {
     isValidationError,
@@ -286,7 +287,8 @@ function toOnCommand<PARAMS>(c: CommandHandlerRegistration<any>): (sdm: MachineO
 }
 
 function toCommandListenerInvocation<P>(c: CommandRegistration<P>, context: HandlerContext, parameters: P): CommandListenerInvocation {
-    let credentials; // opts.credentialsResolver.commandHandlerCredentials(context, undefined);
+    // It may already be there
+    let credentials = !!context ? (context as any as SdmContext).credentials : undefined;
     let ids: RemoteRepoRef[];
     if (isSeedDrivenGeneratorParameters(parameters)) {
         credentials = parameters.target.credentials;