+++ test/api-helper/listener/executeAutoInspect.test.ts
@@ -33,6 +33,7 @@ import {
     ReviewListener,
     ReviewListenerInvocation,
 } from "../../../lib/api/listener/ReviewListener";
+import { AutoInspectRegistration } from "../../../lib/api/registration/AutoInspectRegistration";
 import { PushImpactResponse } from "../../../lib/api/registration/PushImpactListenerRegistration";
 
 const HatesTheWorld: ReviewerRegistration = {
@@ -49,7 +50,6 @@ const HatesTheWorld: ReviewerRegistration = {
                     offset: -1,
                 })),
     }),
-    options: { considerOnlyChangedFiles: false },
 };
 
 const JustTheOne: ReviewerRegistration = {
@@ -66,7 +66,6 @@ const JustTheOne: ReviewerRegistration = {
                     offset: -1,
                 })],
     }),
-    options: { considerOnlyChangedFiles: false },
 };
 
 function loggingReviewListenerWithApproval(saveTo: ReviewListenerInvocation[]): ReviewListener {
@@ -122,6 +121,31 @@ describe("executeAutoInspects", () => {
         assert.equal(reviewEvents[0].addressChannels, rwlc.addressChannels);
         assert.equal(r.code, 0);
         assert(r.state);
+    });
+
+    it("should find push", async () => {
+        const id = new GitHubRepoRef("a", "b");
+        const p = InMemoryProject.from(id, new InMemoryProjectFile("thing", "1"));
+        const rwlc = fakeGoalInvocation(id, {
+            projectLoader: new SingleProjectLoader(p),
+        } as any);
+        const errors: string[] = [];
+        const reg: AutoInspectRegistration<void> = {
+            name: "reg",
+            inspection: async (project, ci) => {
+                if (project !== p) {
+                    errors.push("Project not the same");
+                }
+                if (!ci.push || ci.push.push !== rwlc.sdmGoal.push) {
+                    errors.push("push should not be set");
+                }
+                assert(!ci);
+            },
+        };
+        const ge = executeAutoInspects([reg], []);
+
+        await ge(rwlc);
+        assert.deepEqual(errors, []);
     });
 
     it("should hate anything it finds, without requiring approval", async () => {