+++ CHANGELOG.md
@@ -42,6 +42,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 -   **BREAKING** Remove SDM-level goal methods. [#545](https://github.com/atomist/sdm/issues/545)
 -   **BREAKING** Remove old style registration methods. [#546](https://github.com/atomist/sdm/issues/546)
 -   **BREAKING** Remove client exports from SDM. [#547](https://github.com/atomist/sdm/issues/547)
+-   **BREAKING** Remove `considerOnlyChangedFiles` setting on autofix and inspection registrations. This should be handled
+by specific registrations.
 
 ### Fixed